requirements.txt
# core app
streamlit>=1.20.0
streamlit-plotly-events==0.0.6
python-dotenv>=1.0.0

# data & viz
pandas>=2.0.0
numpy>=1.22.0
plotly>=5.0.0

# modeling & clustering
scikit-learn>=1.0.0
prophet>=1.1.0

# database
SQLAlchemy>=2.0.0
pymssql>=2.2.7

# misc
requests>=2.28.0
python-dateutil>=2.8.0
